fix: Change so we do not try to logout upon 401 response when logoutOnInvalidtoken is set if the user is not authenticated
Before the change, the logout function was called when there was a 401 response to a login attempt (which is a normal
response for invalid credentials) and logoutOnInvalidtoken was set. The logout() call was not needed and as a side effect
it also hid the response from the rejected promise returned by login(). This made proper error handling hard for login dialogs.

After discussing with @RWOverdijk we decided that the easiest way to fix it was to not attempt to handle the 401 response
at all when the user is not logged in.

Fixes #354